fix: lsp salsa panic
